#173 [refact] new rules data class 데이터 모듈로 이동
