+++ app/src/main/java/com/hous/hous_aos/ui/newrules/component/CheckBox.kt
@@ -22,6 +22,7 @@ import androidx.compose.ui.text.font.Font
 import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.text.font.FontStyle
 import androidx.compose.ui.unit.dp
+import com.hous.data.entity.State
 import com.hous.hous_aos.R
 
 @Composable
