+++ app/src/main/java/com/hous/hous_aos/ui/newrules/component/DeleteButton.kt
@@ -5,6 +5,7 @@ import androidx.compose.foundation.clickable
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
+import com.hous.data.entity.State
 import com.hous.hous_aos.R
 
 @Composable
