+++ app/src/main/java/com/hous/hous_aos/ui/newrules/component/ManagerDropDown.kt
@@ -31,6 +31,7 @@ import androidx.compose.ui.text.font.FontStyle
 import androidx.compose.ui.unit.dp
 import com.hous.hous_aos.R
 import com.hous.data.entity.Homie
+import com.hous.data.entity.State
 import com.hous.hous_aos.ui.newrules.isAddDay
 
 @Composable
