+++ app/src/main/java/com/hous/hous_aos/ui/newrules/component/NewRuleToolBar.kt
@@ -14,6 +14,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.Font
 import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.text.font.FontStyle
+import com.hous.data.entity.State
 import com.hous.hous_aos.R
 
 @Composable
