+++ data/src/main/java/com/hous/data/entity/DayData.kt
@@ -1,0 +1,6 @@
+package com.hous.data.entity
+
+data class DayData(
+    val day: String,
+    val dayState: State
+)
