+++ data/src/main/java/com/hous/data/entity/State.kt
@@ -1,4 +1,4 @@
-package com.hous.hous_aos.ui.newrules.component
+package com.hous.data.entity
 
 enum class State {
     UNSELECT, SELECT, BLOCK
